fix.backend: skip empty files

+++ src/main/java/com/fosskeeters/rrss/controllers/ProductController.java
@@ -149,6 +149,8 @@ public class ProductController {
             int displayOrder = 1;
 
             for (MultipartFile image : productDto.getImages()) {
+                if (image.isEmpty()) continue;
+
                 String storageFileName = LocalDateTime.now() + "_" + image.getOriginalFilename();
                 String storagePathStr = "/product-images/" + storageFileName;
 
@@ -232,6 +234,8 @@ public class ProductController {
             int displayOrder = 1;
 
             for (MultipartFile image : productDto.getImages()) {
+                if (image.isEmpty()) continue;
+
                 String storageFileName = LocalDateTime.now() + "_" + image.getOriginalFilename();
                 String storagePathStr = "/product-images/" + storageFileName;
 
@@ -246,8 +250,10 @@ public class ProductController {
                 displayOrder++;
             }
 
-            productImageRepository.deleteAllByProduct(product.get());
-            productImageRepository.saveAll(productImages);
+            if (!productImages.isEmpty()) {
+                productImageRepository.deleteAllByProduct(product.get());
+                productImageRepository.saveAll(productImages);
+            }
         }
 
         product.get().setFromProductDto(productDto);
